优化数据加载和图像处理逻辑
- 修改 `RawVolumeDataManager::LoadData` 中的总体素计算方式，使用 `static_cast` 提升类型安全性，并添加注释增强可读性。
- 修复 `TiffVolumeDataManager::LoadData` 的异常处理，确保能捕获并输出错误信息。
- 在项目文件中添加 `vtkImagingCore-9.4.lib` 作为依赖项，确保正确链接库。
- 在 `VisualStrategies.cpp` 中引入 `vtkImageResample.h`，并实现降采样逻辑以优化性能。
- 更新 `VolumeStrategy::UpdateVisuals` 方法中的标志条件，确保材质更新时也能响应传递函数。
- 修复 `ColoredPlanesStrategy` 中平面法线方向的设置，提升代码清晰性。
- 修改 `main.cpp` 中的数据加载路径，确保程序能正确加载新数据，并调整交互模式和传递函数设置以改善用户体验。

+++ MVVCVTK/DataManager.cpp
@@ -38,7 +38,7 @@ bool RawVolumeDataManager::LoadData(const std::string& filePath) {
     newImage->AllocateScalars(VTK_FLOAT, 1);
 
     // 读取数据到新内存
-    size_t totalVoxels = (size_t)newDims[0] * newDims[1] * newDims[2];
+    size_t totalVoxels = static_cast<size_t>(newDims[0]) * static_cast<size_t>(newDims[1]) * static_cast<size_t>(newDims[2]);
     float* vtkDataPtr = static_cast<float*>(newImage->GetScalarPointer());
 
     std::ifstream file(filePath, std::ios::binary);
@@ -47,7 +47,7 @@ bool RawVolumeDataManager::LoadData(const std::string& filePath) {
     file.close();
 
     newImage->Modified();
-
+ 
     // --- 提交  ---
     {
         std::lock_guard<std::mutex> lock(m_mutex);
@@ -141,7 +141,7 @@ bool TiffVolumeDataManager::LoadData(const std::string& inputPath) {
             }
             // 如果一个字符串是另一个的前缀，短的在前
             return s1.size() < s2.size();
-            };
+        };
 
 
         // 排序
@@ -171,7 +171,7 @@ bool TiffVolumeDataManager::LoadData(const std::string& inputPath) {
     // 如果是序列图片，Reader 会根据文件数量自动设置 Z 轴维度
     try {
         reader->Update();
-    }
+    }   
     catch (...) {
         std::cerr << "[Error] Exception during TIFF reading." << std::endl;
         return false;
